feat: add interval refresh on the sks people counter (#415)
* feat: add interval refresh on the sks people counter

* refactor: add separate files for refresh on ref and consts

add files: sks_people_live_consts.dart
ref_extensions.dart

change files:
home_viev.dart: add SksAppBar to AppBar

parkings_repository.dart, latest_sks_user_data_repository.dart: refactor code

* refactor: delete unused imports
