+++ lib/features/parkings_view/repository/parkings_repository.dart
@@ -5,22 +5,13 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
+import "../../../utils/ref_extensions.dart";
 import "../api_client/iparking_client.dart";
 import "../api_client/iparking_commands.dart";
 import "../models/parking.dart";
 import "local_fav_parking_repository.dart";
 
 part "parkings_repository.g.dart";
-
-extension RefIntervalRefreshX on Ref {
-  void setRefresh(Duration interval) {
-    final timer = Timer.periodic(
-      interval,
-      (t) => invalidateSelf(),
-    );
-    onDispose(timer.cancel);
-  }
-}
 
 @riverpod
 Future<IList<Parking>> parkingsRepository(Ref ref) async {
