+++ lib/features/sks_people_live/data/models/sks_user_data.dart
@@ -6,13 +6,14 @@ part "sks_user_data.g.dart";
 @freezed
 class SksUserData with _$SksUserData {
   const factory SksUserData({
-    required int id,
     required int activeUsers,
     required int movingAverage21,
     required DateTime externalTimestamp,
     required DateTime createdAt,
     required DateTime updatedAt,
     required Trend trend,
+    required bool isResultRecent,
+    required DateTime nextUpdateTimestamp,
   }) = _SksUserData;
 
   factory SksUserData.fromJson(Map<String, dynamic> json) =>
