+++ lib/features/sks_people_live/data/sks_people_live_consts.dart
@@ -1,0 +1,3 @@
+abstract class SksPeopleRepositoryConfig {
+  static const defaultSksRefreshTime = Duration(minutes: 1);
+}
